anilist-cli-import/requirements.txt
gql[all]
python-dotenv
